tests/wpt/tests/html/canvas/offscreen/shadows/2d.shadow.image.alpha.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.shadow.image.alpha</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.shadow.image.alpha</h1>
<p class="desc">Shadows are drawn correctly for partially-transparent images</p>


<script>
promise_test(async t => {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  ctx.shadowOffsetY = 50;
  ctx.shadowColor = '#00f';
  var response = await fetch('/images/transparent50.png')
  var blob = await response.blob();
  var img = await createImageBitmap(blob);
  ctx.drawImage(img, 0, -50);

  _assertPixelApprox(canvas, 50,25, 127,0,127,255, 2);

}, "Shadows are drawn correctly for partially-transparent images");
</script>
